showroom/usecase-042/requirements.txt
flask==2.3.3
openai>=1.12.0
python-dotenv==1.0.0
Flask-WTF==1.2.1
Werkzeug==2.3.7
jinja2==3.1.2
requests==2.31.0
pydantic==2.5.2